+++ Magic/src/main/java/com/elmakers/mine/bukkit/protection/WorldGuardAPI.java
@@ -20,26 +20,47 @@ import java.util.Set;
 import java.util.logging.Level;
 
 public class WorldGuardAPI {
+    private final Plugin owningPlugin;
 	private WorldGuardPlugin worldGuard = null;
-    private WorldGuardFlagsManager customFlags = null;
+    private WorldGuardFlags customFlags = null;
 
 	public boolean isEnabled() {
 		return worldGuard != null;
 	}
 	
 	public WorldGuardAPI(Plugin plugin, Plugin owningPlugin) {
+        this.owningPlugin = owningPlugin;
         if (plugin instanceof WorldGuardPlugin) {
             worldGuard = (WorldGuardPlugin)plugin;
             try {
                 owningPlugin.getLogger().info("Pre-check for WorldGuard custom flag registration");
                 customFlags = new WorldGuardFlagsManager(owningPlugin, worldGuard);
             } catch (NoSuchMethodError incompatible) {
-                owningPlugin.getLogger().log(Level.WARNING, "Failed to set up custom flags, please make sure you are on WorldGuard 6.2 or above");
+                // Ignored, will follow up in checkFlagSupport
             } catch (Throwable ex) {
                 owningPlugin.getLogger().log(Level.WARNING, "Unexpected error setting up custom flags, please make sure you are on WorldGuard 6.2 or above", ex);
             }
         }
 	}
+	
+	public void checkFlagSupport() {
+        if (customFlags == null) {
+            try {
+                Plugin customFlagsPlugin = owningPlugin.getServer().getPluginManager().getPlugin("WGCustomFlags");
+                if (customFlagsPlugin != null) {
+                    customFlags = new WGCustomFlagsManager(customFlagsPlugin);
+                }
+            } catch (Throwable ex) {
+                owningPlugin.getLogger().log(Level.WARNING, "Error integration with WGCustomFlags", ex);
+            }
+
+            if (customFlags != null) {
+                owningPlugin.getLogger().info("WGCustomFlags found, added custom flags");
+            } else {
+                owningPlugin.getLogger().log(Level.WARNING, "Failed to set up custom flags, please make sure you are on WorldGuard 6.2 or above, or use the WGCustomFlags plugin");
+            }
+        }
+    }
 
     protected RegionAssociable getAssociable(Player player) {
         RegionAssociable associable;