+++ Magic/src/main/java/com/elmakers/mine/bukkit/protection/WorldGuardFlagsManager.java
@@ -14,7 +14,7 @@ import org.bukkit.plugin.Plugin;
 
 import java.util.Set;
 
-public class WorldGuardFlagsManager {
+public class WorldGuardFlagsManager implements WorldGuardFlags {
 
     private final WorldGuardPlugin customFlags;